BUG: Fix build
