+++ single-node-refactor/src/CMakeLists.txt
@@ -75,5 +75,9 @@ endif()
 
 include_directories(common)
 
-add_executable(Fierro main.cpp solver.cpp )
-target_link_libraries(Fierro PRIVATE matar parse_yaml sgh_solver Kokkos::kokkos)
+# Add SGH Solver
+include_directories(Solvers/SGH_solver/include)
+add_subdirectory(Solvers/SGH_solver)
+
+add_executable(Fierro main.cpp solver.cpp ${YAML_SRC_Files} ${SGH_SRC_Files} )
+target_link_libraries(Fierro PRIVATE matar Kokkos::kokkos) #sgh_solver  parse_yaml
