+++ single-node-refactor/src/Solvers/SGH_solver/include/sgh_solver.h
@@ -47,8 +47,6 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "dynamic_options.h"
 
 
-
-
 using namespace mtr; // matar namespace
 
 /////////////////////////////////////////////////////////////////////////////
