+++ single-node-refactor/src/common/region_fill.h
@@ -40,12 +40,342 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "state.h"
 #include "simulation_parameters.h"
 #include "region.h"
+#include "io_utils.h"
 
 #include <stdio.h>
 
+// ==============================================================================
+//   Functions to read voxel mesh
+// ==============================================================================
+void user_voxel_init(DCArrayKokkos<size_t>& elem_values,
+                     double& dx,
+                     double& dy,
+                     double& dz,
+                     double& orig_x,
+                     double& orig_y,
+                     double& orig_z,
+                     size_t& voxel_num_i,
+                     size_t& voxel_num_j,
+                     size_t& voxel_num_k,
+                     double scale_x,
+                     double scale_y,
+                     double scale_z,
+                     std::string mesh_file);
 
+// -----------------------------------------------------------------------------
+// The function to read a voxel vtk file from Dream3d and intialize the mesh
+// ------------------------------------------------------------------------------
+void user_voxel_init(DCArrayKokkos<size_t>& elem_values,
+                     double& dx,
+                     double& dy,
+                     double& dz,
+                     double& orig_x,
+                     double& orig_y,
+                     double& orig_z,
+                     size_t& num_elems_i,
+                     size_t& num_elems_j,
+                     size_t& num_elems_k,
+                     double scale_x,
+                     double scale_y,
+                     double scale_z,
+                     std::string mesh_file)
+{
+    std::string MESH = mesh_file; // user specified
 
+    std::ifstream in;  // FILE *in;
+    in.open(MESH);
 
+    // check to see of a mesh was supplied when running the code
+    if (in)
+    {
+        printf("\nReading the 3D voxel mesh: ");
+        std::cout << MESH << std::endl;
+    }
+    else
+    {
+        std::cout << "\n\n**********************************\n\n";
+        std::cout << " ERROR:\n";
+        std::cout << " Voxel vtk input does not exist \n";
+        std::cout << "**********************************\n\n" << std::endl;
+        std::exit(EXIT_FAILURE);
+    } // end if
+
+    size_t i;           // used for writing information to file
+    size_t point_id;    // the global id for the point
+    size_t elem_id;     // the global id for the elem
+    size_t this_point;   // a local id for a point in a elem (0:7 for a Hexahedral elem)
+
+    size_t num_points_i;
+    size_t num_points_j;
+    size_t num_points_k;
+
+    size_t num_dims = 3;
+
+    std::string token;
+
+    bool found = false;
+
+    // look for POINTS
+    i = 0;
+    while (found == false) {
+        std::string str;
+        std::string delimiter = " ";
+        std::getline(in, str);
+        std::vector<std::string> v = split(str, delimiter);
+
+        // looking for the following text:
+        //      POINTS %d float
+        if (v[0] == "DIMENSIONS")
+        {
+            num_points_i = std::stoi(v[1]);
+            num_points_j = std::stoi(v[2]);
+            num_points_k = std::stoi(v[3]);
+            printf("Num voxel nodes read in = %zu, %zu, %zu\n", num_points_i, num_points_j, num_points_k);
+
+            found = true;
+        } // end if
+
+        if (i > 1000)
+        {
+            printf("ERROR: Failed to find POINTS \n");
+            break;
+        } // end if
+
+        i++;
+    } // end while
+
+    found = false;
+
+    int num_points = num_points_i * num_points_j * num_points_k;
+    CArray<double> pt_coords_x(num_points_i);
+    CArray<double> pt_coords_y(num_points_j);
+    CArray<double> pt_coords_z(num_points_k);
+
+    while (found == false) {
+        std::string str;
+        std::string str0;
+        std::string delimiter = " ";
+        std::getline(in, str);
+        std::vector<std::string> v = split(str, delimiter);
+
+        // looking for the following text:
+        if (v[0] == "X_COORDINATES")
+        {
+            size_t num_saved = 0;
+
+            while (num_saved < num_points_i - 1) {
+                // get next line
+                std::getline(in, str0);
+
+                // remove starting and trailing spaces
+                str = trim(str0);
+                std::vector<std::string> v_coords = split(str, delimiter);
+
+                // loop over the contents of the vector v_coords
+                for (size_t this_point = 0; this_point < v_coords.size(); this_point++)
+                {
+                    pt_coords_x(num_saved) = scale_x*std::stod(v_coords[this_point]);
+                    num_saved++;
+                } // end for
+            } // end while
+
+            found = true;
+        } // end if
+
+        if (i > 1000)
+        {
+            printf("ERROR: Failed to find X_COORDINATES \n");
+            break;
+        } // end if
+
+        i++;
+    } // end while
+    found = false;
+
+    while (found == false) {
+        std::string str;
+        std::string str0;
+        std::string delimiter = " ";
+        std::getline(in, str);
+        std::vector<std::string> v = split(str, delimiter);
+
+        // looking for the following text:
+        if (v[0] == "Y_COORDINATES")
+        {
+            size_t num_saved = 0;
+
+            while (num_saved < num_points_j - 1) {
+                // get next line
+                std::getline(in, str0);
+
+                // remove starting and trailing spaces
+                str = trim(str0);
+                std::vector<std::string> v_coords = split(str, delimiter);
+
+                // loop over the contents of the vector v_coords
+                for (size_t this_point = 0; this_point < v_coords.size(); this_point++)
+                {
+                    pt_coords_y(num_saved) = scale_y*std::stod(v_coords[this_point]);
+                    num_saved++;
+                } // end for
+            } // end while
+
+            found = true;
+        } // end if
+
+        if (i > 1000)
+        {
+            printf("ERROR: Failed to find Y_COORDINATES \n");
+            break;
+        } // end if
+
+        i++;
+    } // end while
+    found = false;
+
+    while (found == false) {
+        std::string str;
+        std::string str0;
+        std::string delimiter = " ";
+        std::getline(in, str);
+        std::vector<std::string> v = split(str, delimiter);
+
+        // looking for the following text:
+        if (v[0] == "Z_COORDINATES")
+        {
+            size_t num_saved = 0;
+
+            while (num_saved < num_points_k - 1) {
+                // get next line
+                std::getline(in, str0);
+
+                // remove starting and trailing spaces
+                str = trim(str0);
+                std::vector<std::string> v_coords = split(str, delimiter);
+
+                // loop over the contents of the vector v_coords
+                for (size_t this_point = 0; this_point < v_coords.size(); this_point++)
+                {
+                    pt_coords_z(num_saved) = scale_z*std::stod(v_coords[this_point]);
+                    num_saved++;
+                } // end for
+            } // end while
+
+            found = true;
+        } // end if
+
+        if (i > 1000)
+        {
+            printf("ERROR: Failed to find Z_COORDINATES \n");
+            break;
+        } // end if
+
+        i++;
+    } // end while
+    found = false;
+
+    size_t num_elems;
+    num_elems_i = num_points_i - 1;
+    num_elems_j = num_points_j - 1;
+    num_elems_k = num_points_k - 1;
+
+    // center to center distance between first and last elem along each edge
+    double Lx = (pt_coords_x(num_points_i - 2) - pt_coords_x(0));
+    double Ly = (pt_coords_y(num_points_j - 2) - pt_coords_y(0));
+    double Lz = (pt_coords_z(num_points_k - 2) - pt_coords_z(0));
+
+    // spacing between elems
+    dx = Lx / ((double) num_elems_i);
+    dy = Ly / ((double) num_elems_j);
+    dz = Lz / ((double) num_elems_k);
+
+    // element mesh origin
+    orig_x = 0.5 * (pt_coords_x(0) + pt_coords_x(1)),
+    orig_y = 0.5 * (pt_coords_y(0) + pt_coords_y(1)),
+    orig_z = 0.5 * (pt_coords_z(0) + pt_coords_z(1)),
+
+    // look for CELLS
+    i = 0;
+    while (found == false) {
+        std::string str;
+        std::getline(in, str);
+
+        std::string              delimiter = " ";
+        std::vector<std::string> v = split(str, delimiter);
+
+        // looking for the following text:
+        //      CELLS num_elems size
+        if (v[0] == "CELL_DATA")
+        {
+            num_elems = std::stoi(v[1]);
+            printf("Num voxel elements read in %zu\n", num_elems);
+
+            found = true;
+        } // end if
+
+        if (i > 1000)
+        {
+            printf("ERROR: Failed to find CELL_DATA \n");
+            break;
+        } // end if
+
+        i++;
+    } // end while
+    found = false;
+
+    // allocate memory for element voxel values
+    elem_values = DCArrayKokkos<size_t>(num_elems);
+
+    // reading the cell data
+    while (found == false) {
+        std::string str;
+        std::string str0;
+
+        std::string delimiter = " ";
+        std::getline(in, str);
+        std::vector<std::string> v = split(str, delimiter);
+
+        // looking for the following text:
+        if (v[0] == "LOOKUP_TABLE")
+        {
+            size_t num_saved = 0;
+
+            while (num_saved < num_elems - 1) {
+                // get next line
+                std::getline(in, str0);
+
+                // remove starting and trailing spaces
+                str = trim(str0);
+                std::vector<std::string> v_values = split(str, delimiter);
+
+                // loop over the contents of the vector v_coords
+                for (size_t this_elem = 0; this_elem < v_values.size(); this_elem++)
+                {
+                    // save integers (0 or 1) to host side
+                    elem_values.host(num_saved) = std::stoi(v_values[this_elem]);
+                    num_saved++;
+                } // end for
+
+                // printf(" done with one row of data \n");
+            } // end while
+
+            found = true;
+        } // end if
+
+        if (i > 1000)
+        {
+            printf("ERROR: Failed to find LOOKUP_TABLE data \n");
+            break;
+        } // end if
+
+        i++;
+    } // end while
+    found = false;
+
+    printf("\n");
+
+    in.close();
+} // end routine
 /////////////////////////////////////////////////////////////////////////////
 ///
 /// \fn fill_geometric_region
