+++ single-node-refactor/src/input/CMakeLists.txt
@@ -18,16 +18,21 @@ endif()
 
 include_directories(../common)
 
-set(SRC_Files 
-parse_yaml.cpp
-Yaml.cpp)
+set(YAML_SRC_Files 
+${CMAKE_CURRENT_SOURCE_DIR}/parse_yaml.cpp
+${CMAKE_CURRENT_SOURCE_DIR}/parse_yaml.h
+${CMAKE_CURRENT_SOURCE_DIR}/Yaml.cpp
+${CMAKE_CURRENT_SOURCE_DIR}/Yaml.hpp
+PARENT_SCOPE
+)
 
 # set(SGH_Solver_SRC src/sgh_solver.cpp )
 
 message("\n ****** ADDING PARSE_YAML LIBRARY ******** \n ")
 
-add_library(parse_yaml ${SRC_Files} ) 
-target_link_libraries(parse_yaml matar Kokkos::kokkos)
+
+# add_library(parse_yaml ${SRC_Files} ) 
+# target_link_libraries(parse_yaml matar Kokkos::kokkos)
 # target_include_directories(sgh_solver PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
