+++ single-node-refactor/src/input/parse_yaml.cpp
@@ -105,6 +105,65 @@ std::vector<std::string> exact_array_values(std::string s, std::string delimiter
     return res;
 } // end of extract_array_values
 
+// Code from stackover flow for string delimiter parsing
+std::vector<std::string> split(std::string s, std::string delimiter)
+{
+    size_t                   pos_start = 0, pos_end, delim_len = delimiter.length();
+    std::string              token;
+    std::vector<std::string> res;
+
+    while ((pos_end = s.find(delimiter, pos_start)) != std::string::npos) {
+        token     = s.substr(pos_start, pos_end - pos_start);
+        pos_start = pos_end + delim_len;
+        res.push_back(token);
+    }
+
+    res.push_back(s.substr(pos_start));
+    return res;
+} // end of split
+
+
+// retrieves multiple values between [ ]
+std::vector<double> extract_list(std::string str)
+{
+    // replace '[' with a space and ']' with a space
+    std::replace(str.begin(), str.end(), '[', ' ');
+    std::replace(str.begin(), str.end(), ']', ' ');
+
+    std::vector<std::string> str_values;
+    std::vector<double>      values;
+
+    // exact the str values into a vector
+    str_values = split(str, ",");
+
+    // convert the text values into double values
+    for (auto& word : str_values)
+    {
+        values.push_back(atof(word.c_str()) );
+    } // end for
+
+    return values;
+}  // end of extract_list
+
+
+std::string ltrim(const std::string& s)
+{
+    size_t start = s.find_first_not_of(WHITESPACE);
+    return (start == std::string::npos) ? "" : s.substr(start);
+}
+
+
+std::string rtrim(const std::string& s)
+{
+    size_t end = s.find_last_not_of(WHITESPACE);
+    return (end == std::string::npos) ? "" : s.substr(0, end + 1);
+}
+
+std::string trim(const std::string& s)
+{
+    return rtrim(ltrim(s));
+}
+
 // =================================================================================
 //    Print a yaml file to 6 levels
 // =================================================================================
