+++ single-node-refactor/src/solver.h
@@ -53,7 +53,7 @@ class Solver
 public:
 
     Solver();
-    virtual ~Solver();
+    ~Solver();
 
     virtual void initialize(SimulationParameters_t& SimulationParamaters, 
                             Material_t& Materials,
